pyproject.toml
[project]
name = "teacher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "pandas>=2.3.3",
    "streamlit>=1.52.2",
]
